+++ core/components/romanesco/elements/snippets/06_formulas/f_formblocks/f_fb_modifier/fbstripasalias.snippet.php
@@ -8,11 +8,6 @@
  * @var string $options
  */
 
-$input = strip_tags($input); // strip HTML
-$input = strtolower($input); // convert to lowercase
-$input = preg_replace('/[^A-Za-z0-9 _-]/', '', $input); // strip non-alphanumeric characters
-$input = preg_replace('/\s+/', '-', $input); // convert white-space to dash
-$input = preg_replace('/-+/', '-', $input);  // convert multiple dashes to one
-$input = trim($input, '-'); // trim excess
-
-return $input;
+return $modx->filterPathSegment($input, [
+    'friendly_alias_restrict_chars' => 'alphanumeric'
+]);